requirements.txt
fastapi
uvicorn
pillow
PyMuPDF
numpy
nsfwlib  # Replace with actual NSFW library like nsfw-detector if needed
